requirements.txt
flask
numpy
opencv-python-headless
pillow
pytesseract
